+++ Lib/fontbakery/checks/opentype/cff.py
@@ -10,6 +10,7 @@ class CFFAnalysis:
         self.glyphs_endchar_seac = []
         self.glyphs_exceed_max = []
         self.glyphs_recursion_errors = []
+        self.string_not_ascii = []
 
 
 def _get_subr_bias(count):
@@ -61,6 +62,14 @@ def _analyze_cff(analysis, top_dict, private_dict, fd_index=0):
     global_subrs = top_dict.GlobalSubrs
     gsubr_bias = _get_subr_bias(len(global_subrs))
 
+    if hasattr(top_dict, "rawDict"):
+        raw_dict = top_dict.rawDict
+        for key in ["Notice", "Copyright", "FontName", "FullName", "FamilyName"]:
+            for char in raw_dict.get(key, ""):
+                if ord(char) > 0x7F:
+                    analysis.string_not_ascii.append((key, raw_dict[key]))
+                    break
+
     if private_dict is not None and hasattr(private_dict, "Subrs"):
         subrs = private_dict.Subrs
         subr_bias = _get_subr_bias(len(subrs))
@@ -109,7 +118,11 @@ def cff_analysis(font):
         ttFont = TTFont(font.file)  # Use our own copy here since we are decompiling
 
     if "CFF " in ttFont:
-        cff = ttFont["CFF "].cff
+        try:
+            cff = ttFont["CFF "].cff
+        except UnicodeDecodeError:
+            analysis.string_not_ascii = None
+            return analysis
 
         for top_dict in cff.topDictIndex:
             if hasattr(top_dict, "FDArray"):
@@ -218,3 +231,35 @@ def com_adobe_fonts_check_cff_deprecated_operators(cff_analysis):
                 f'Glyph "{gn}" has deprecated use of "endchar"'
                 f" operator to build accented characters (seac).",
             )
+
+
+@check(
+    id="com.adobe.fonts/check/cff_ascii_strings",
+    conditions=["ttFont", "is_cff", "cff_analysis"],
+    rationale="""
+        All CFF Table top dict string chars should fit into the ASCII range.
+    """,
+    proposal="https://github.com/fonttools/fontbakery/issues/4619",
+    experimental="Since 2024/Jun/20",
+)
+def com_adobe_fonts_check_cff_ascii_strings(cff_analysis):
+    """Does the font's CFF table top dict strings fit into the ASCII range?"""
+
+    if cff_analysis.string_not_ascii is None:
+        yield FAIL, Message(
+            "cff-unable-to-decode",
+            "Unable to decode CFF table, possibly due to out"
+            " of ASCII range strings. Please check table strings.",
+        )
+    elif cff_analysis.string_not_ascii:
+        detailed_info = ""
+        for key, string in cff_analysis.string_not_ascii:
+            detailed_info += (
+                f"\n\n\t - {key}: {string.encode('latin-1').decode('utf-8')}"
+            )
+
+        yield FAIL, Message(
+            "cff-string-not-in-ascii-range",
+            f"The following CFF TopDict strings"
+            f" are not in the ASCII range: {detailed_info}",
+        )